zram: rework copy of compressor name in comp_algorithm_store()
comp_algorithm_store() passes the size of the source buffer to strlcpy()
instead of the destination buffer size.  Make it explicit that the two buffers
have the same size and use strcpy() instead of strlcpy().  The latter can be
done safely since the function ensures that the string in the source buffer is
terminated.

Link: http://lkml.kernel.org/r/20170803163350.45245-1-mka@chromium.org

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>
Acked-by: Minchan Kim <minchan@kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: utsavbalar1231 <utsavbalar1231@gmail.com>